Add description
The above file uses the computer dataset and my function_0822.R. Two ways of collapsing small groups are included, Helmert Coding and Dummy Coding. Below is a small description of the file.
Helmert Coding: 
Levels of the factor where small-group issue exists are first sorted in an ascending order based on their number of instances. Then Helmert Coding is applied on all factors and a lasso model is trained. Based on model coefs, that code block outputs the first level whose coef is not zero of each factor. In other words, levels that are sorted before the output level need to be collapsed to that output level. It is worth to mention that we can get the result in one iteration, but the result is not very stable. 
Dummy Coding:
This code block can only work on one factor at one iteration, so firstly a factor where small-group issue exists needs to be chosen. Levels of that factor are first sorted in an ascending order based on their number of instances. According to how we define a group is "small", we select levels that need to be collapsed. Then for each of the level to be collapsed `l`, we select it as the reference group of dummy coding and run a lasso model with all other factors in dummy codings. According to model coefs, the first level whose coef equals to zero is the level that `l` should be assigned to. It is worth to mention that based on how I write this code block, for a factor, we need to run one model for each small group and then repeat the process for each factor. Therefore, more iterations are needed, but the result is more stable.